Optimized tex2js conversion script; fixed broken exceptions for Bahasa Indonesia patterns
