+++ CHANGELOG
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Optimized tex2js conversion script
+- Fixed broken exceptions for Bahasa Indonesia patterns
+
 ## [1.10.6]
 
 ### Changed
