+++ patterns/id.js
@@ -16,7 +16,12 @@
     "01,21,201,232,0023,00203,2,021,02,10102,3,0202,00304",
     '{"a":[{"i":{"r":8},"n":{".":10}},0],"e":0,"i":[{"o":{"n":7}},0],"o":0,"u":0,"b":{"d":1,"j":1,"k":1,"n":1,"s":1,"t":1,"a":{"g":{"a":{"i":9}},"n":{".":1}}},"c":{"k":1,"n":1,"a":{"n":{".":1}}},"d":{"k":1,"n":1,"p":1,"a":{"n":{".":1}}},"f":{"d":1,"k":1,"n":1,"t":1,"a":{"n":{".":1}}},"g":{"g":1,"k":1,"n":1,"a":{"n":{".":1}}},"h":{"k":1,"l":1,"m":1,"n":1,"w":1,"a":{"n":{".":1}}},"j":{"k":1,"n":1,"a":{"n":{".":1}}},"k":{"b":1,"k":1,"m":1,"n":1,"r":1,"s":1,"t":1,"a":{"n":{".":1}}},"l":{"b":1,"f":1,"g":1,"h":1,"k":1,"m":1,"n":1,"s":1,"t":1,"q":1,"a":{"n":{".":1}}},"m":{"b":1,"k":1,"l":1,"m":1,"n":1,"p":1,"r":1,"s":1,"a":{"n":{".":1}}},"n":{"c":1,"d":1,"f":1,"j":1,"k":1,"n":1,"p":1,"s":[{"t":3},1],"t":1,"v":1,"g":{"g":2,"h":2,"k":2,"n":2,"s":2,".":6,"a":{"n":{".":2}}},"y":{".":6},"a":{"n":{".":1}}},"p":{"k":1,"n":1,"p":1,"r":1,"t":1,"a":{"n":{".":1}}},"r":{"b":1,"c":1,"f":1,"g":1,"h":1,"j":1,"k":1,"l":1,"m":1,"n":1,"p":1,"r":1,"s":1,"t":1,"w":1,"y":1,"a":{"n":{".":1}}},"s":{"b":1,"k":1,"l":1,"m":1,"n":1,"p":1,"r":1,"s":1,"t":1,"w":1,"a":{"n":{".":1}}},"t":{"k":1,"l":1,"n":1,"t":1,"a":{"n":{".":1}}},"w":{"t":1},".":{"b":{"e":{"r":4}},"t":{"e":{"r":4},"a":{"n":{"g":{"a":{"n":{".":12}}}}}},"m":{"e":{"n":{"g":5}},"a":{"n":{"g":{"a":{"n":{".":12}}}}}},"p":{"e":{"r":4},"a":{"n":{"g":{"a":{"n":{".":12}}}}}},"a":{"t":{"a":{"u":11}}},"l":{"e":{"n":{"g":{"a":{"n":{".":12}}}}}},"j":{"a":{"n":{"g":{"a":{"n":{".":12}}}}}},"r":{"i":{"n":{"g":{"a":{"n":{".":12}}}}}},"d":{"e":{"n":{"g":{"a":{"n":{".":12}}}}}}},"v":{"a":{"n":{".":1}}},"z":{"a":{"n":{".":1}}}}',
     [
-      "be-ra-be be-ra-hi be-rak be-ran-da be-ran-dal be-rang",
+      "be-ra-be",
+      "be-ra-hi",
+      "be-rak",
+      "be-ran-da",
+      "be-ran-dal",
+      "be-rang",
       "be-ra-ngas-an",
       "be-rang-sang",
       "be-ra-ngus",
@@ -61,7 +66,11 @@
       "te-rom-pah",
       "te-rom-pet",
       "te-ro-pong",
-      "             te-ro-wong-an te-ru-buk te-ru-na te-rus te-ru-si"
+      "te-ro-wong-an",
+      "te-ru-buk",
+      "te-ru-na",
+      "te-rus",
+      "te-ru-si"
     ]
   ];
 });
